Bad antennas are identified and flagged in step 1
The calculation of the reference antenna has been moved from parallel_cal.py to
validate_input.py, and outlier antennas are passed into flag_round_1.py so that
they are flagged prior to any calibration. The bad antennas are written into the
config file in the list bad_ants=[]. The default config has been updated to
reflect this.