partition now modifies the config file
If partition runs, it modifies the config file to overwrite the `vis`
parameter to hold the name of the MMS, and adds a new keyword `orig_vis`
that has the name of the original visibility.

The other scripts have been modified to read whatever is in the `vis`
keyword, and not do any local replacing of `.ms` -> `.mms`. This makes
it easier to skip partition by simply omitting the call from the config
file.