bugfixes to read_ms.py and xy_yx_solve
The check_refant function always treats the input antenna as a string,
even if it's a number. This causes problems for other telescopes like
EVLA. This has been fixed to first check if the antenna can be converted
to an int, and if not consider it a string.

Assorted bugfixes to xy_yx_solve.